Added pdf and .txt versions of book.
The .txt version has a lot of errors; I got it from the default `Save as other / ...Text` menu item in Acrobat. An automated tool could rejoin the lines that end in hyphens, and perhaps find missing spaces, as in `programmingpractices` and `anunfortunate`. Other errors would require significant human labor to clean up.