Update chapter4.md
Corrected many markdown issues especially with regards to comma separated lists. 

Also fixed an error that appears in the main code for this chapter. The function "op" is not defined by is used as a short-hand for make-op, by using "op" the :action keyword is left out entirely in many subsequent definitions of operation lists.

I have yet to transform pictures to tables, I won't be doing that this session. I may still give this chapter another look through, there are some very hard to find markdown issues like the one I found with "some" being used in achieve-all. Also note that I changed a code reference. Outside a block of code achieve-all was said to be making changes to state2, that is not true, it was updating current-state.