Fix integrating constants and unary minus
Integrating constants was broken for integration variables that were not
called "x".  This mistake was likely due to coincidentally calling the
integration variable "x" in the code, and forgetting to unquote it in
the output expression.  The book only tests cases where "x" is the
integration variable.

Additionally, integrating expressions with a leading "-" silently
dropped the "-".  The book never tests this case.